+++ packages/filterlist/src/types.ts
@@ -167,6 +167,12 @@ export type Params<Item, Additional, Error> = Readonly<{
 	 */
 	pageSize?: number | null;
 	/**
+	 * check whether the list should be refreshed by timeout
+	 * 
+	 * @example shouldRefresh: () => !document.hidden,
+	 */
+	shouldRefresh?: () => boolean;
+	/**
 	 * timeout to refresh the list
 	 */
 	refreshTimeout?: number;